Move dag_algo to its own module
